# src/album/entities/album.entity.ts
import { Artist } from '../../artist/entities/artist.entity';
import {
  Column,
  Entity,
  JoinColumn,
  ManyToOne,
  PrimaryGeneratedColumn,
} from 'typeorm';

export type AlbumId = string;

@Entity()
export class Album {
  @PrimaryGeneratedColumn('uuid')
  id: AlbumId; // uuid v4

  @Column()
  name: string;

  @Column()
  year: number;

  @ManyToOne(() => Artist, (artist) => artist.id, {
    nullable: true,
    onDelete: 'SET NULL',
  })
  @JoinColumn({ name: 'artistId' })
  artist: Artist | null; // refers to Artist

  @Column({
    default: false,
  })
  isFavorite?: boolean;
}