# src/user/user.controller.ts
import {
  Controller,
  Get,
  Post,
  Body,
  Param,
  Delete,
  Put,
  HttpCode,
} from '@nestjs/common';
import { UserService } from './user.service';
import { CreateUserDto } from './dto/create-user.dto';
import { UpdateUserDto } from './dto/update-user.dto';
import { UserId } from './entities/user.entity';
import prepareUserForResponse from './utils/prepare-user-for-response.util';
import uuidValidateV4 from './utils/uuid-user-validate.util';
import IncorrectIdError from './errors/incorrect-user-id.error copy';
import UserNotFoundError from './errors/user-not-found.error';
import InvalidUserPassword from './errors/invalid-user-password.error';

@Controller('user')
export class UserController {
  constructor(private readonly userService: UserService) {}

  @Post()
  async create(@Body() createUserDto: CreateUserDto) {
    const newUser = await this.userService.create(createUserDto);
    return prepareUserForResponse(newUser);
  }

  @Get()
  async findAll() {
    const allUsers = await this.userService.findAll();
    return allUsers.map((user) => prepareUserForResponse(user));
  }

  @Get(':id')
  async findOne(@Param('id') id: UserId) {
    if (!uuidValidateV4(id)) throw new IncorrectIdError();
    const user = await this.userService.findOne(id);
    if (!user) throw new UserNotFoundError();
    return prepareUserForResponse(user);
  }

  @Put(':id')
  async update(@Param('id') id: UserId, @Body() updateUserDto: UpdateUserDto) {
    if (!uuidValidateV4(id)) throw new IncorrectIdError();
    const user = await this.userService.findOne(id);
    if (!user) throw new UserNotFoundError();
    if (user.password !== updateUserDto.oldPassword) {
      throw new InvalidUserPassword();
    }
    const updatedUser = await this.userService.update(id, updateUserDto);
    return prepareUserForResponse(updatedUser);
  }

  @Delete(':id')
  @HttpCode(204)
  async remove(@Param('id') id: UserId) {
    if (!uuidValidateV4(id)) throw new IncorrectIdError();
    const isDeleted = await this.userService.remove(id);
    if (!isDeleted) throw new UserNotFoundError();
    return;
  }
}